Use UTC timestamp and do actual test in integration test
